[frontend] Audits List error fix
